test: implement dex swap on mobile
