+++ apps/ledger-live-mobile/src/screens/Swap/LiveApp/navigationHandlers/useSwapHeaderNavigation.tsx
@@ -124,7 +124,7 @@ export function useSwapHeaderNavigation(webviewRef: React.RefObject<WebviewAPI |
     if (isTwoStepApproval && isTransactionComplete) {
       navigation.setOptions({
         headerLeft: undefined,
-        headerRight: () => <NavigationHeaderCloseButton onPress={navigateToSwapForm} />,
+        headerRight: () => <NavigationHeaderCloseButton testIDSuffix="Completed" onPress={navigateToSwapForm} />,
         headerTitle: getScreenTitle({ t, webviewCurrentPage, isTransactionComplete }),
       });
     } else if (
