+++ e2e/desktop/tests/utils/cliCommandsUtils.ts
@@ -1,10 +1,12 @@
 import invariant from "invariant";
-import { parseCurrencyUnit } from "@ledgerhq/live-common/currencies/index";
-import { CLI } from "tests/utils/cliUtils";
+import { CLI, GetTokenAllowanceOpts, TokenApprovalOpts } from "tests/utils/cliUtils";
 import { Account, TokenAccount } from "@ledgerhq/live-common/e2e/enum/Account";
 import { Currency } from "@ledgerhq/live-common/e2e/enum/Currency";
 import { Transaction } from "@ledgerhq/live-common/e2e/models/Transaction";
-import { approveToken } from "@ledgerhq/live-common/e2e/families/evm";
+import {
+  createApproveTokenCommand,
+  createIsTokenAllowanceSufficientCommand,
+} from "@ledgerhq/live-common/e2e";
 
 type LiveDataCommandOptions = {
   readonly useScheme?: boolean;
@@ -95,99 +97,10 @@ export const liveDataWithRecipientAddressCommand = (
     return address;
   };
 };
-export const approveTokenCommand =
-  (account: TokenAccount, spender: string, approveAmount: string) => async () => {
-    const originalDisableBroadcast = setDisableBroadcast("0");
+export const approveTokenCommand = createApproveTokenCommand((opts: TokenApprovalOpts) =>
+  CLI.tokenApproval(opts),
+);
 
-    const result = CLI.tokenApproval({
-      currency: account.currency.speculosApp.name,
-      index: account.index,
-      spender: spender,
-      token: account.currency.id,
-      mode: "approve",
-      approveAmount,
-      waitConfirmation: true,
-    });
-
-    try {
-      await approveToken();
-    } finally {
-      restoreDisableBroadcast(originalDisableBroadcast);
-    }
-    return await result;
-  };
-
-export const isTokenAllowanceSufficientCommand =
-  (account: TokenAccount, spenderAddress: string, minAmount: string) => async () => {
-    const ownerAddress = account.parentAccount?.address ?? account.address;
-    if (!ownerAddress) throw new Error("Token allowance check requires the main account address");
-
-    const output = await CLI.getTokenAllowance({
-      currency: account.currency.speculosApp.name,
-      token: account.currency.id,
-      spenderAddress,
-      index: account.index,
-      format: "json",
-      ownerAddress,
-    });
-
-    const { allowanceStr, unitMagnitude } = parseTokenAllowanceCliOutput(output);
-
-    const smallestUnit = { name: "smallest", code: "", magnitude: unitMagnitude } as const;
-    const minInSmallestUnit = parseCurrencyUnit(smallestUnit, minAmount);
-    const minStr = minInSmallestUnit.toFixed(0);
-
-    const allowanceBi = BigInt(allowanceStr);
-    const minBi = BigInt(minStr);
-    if (allowanceBi >= minBi) return allowanceStr;
-    return 0;
-  };
-
-function parseTokenAllowanceCliOutput(output: string): {
-  allowanceStr: string;
-  unitMagnitude: number;
-} {
-  const jsonStart = output.indexOf("{");
-  if (jsonStart === -1) throw new Error("No JSON found in tokenAllowance output:\n" + output);
-
-  const rawParsed: unknown = JSON.parse(output.slice(jsonStart));
-  if (typeof rawParsed !== "object" || rawParsed === null) {
-    throw new Error("Invalid tokenAllowance JSON:\n" + output);
-  }
-
-  const allowanceField = Reflect.get(rawParsed, "allowance");
-  if (typeof allowanceField !== "string") {
-    throw new Error("Invalid tokenAllowance JSON (allowance):\n" + output);
-  }
-  const allowanceStr = allowanceField.trim();
-  if (!/^\d+$/.test(allowanceStr)) {
-    throw new Error("Invalid raw allowance in tokenAllowance:\n" + output);
-  }
-
-  const magnitudeField = Reflect.get(rawParsed, "unitMagnitude");
-  if (
-    typeof magnitudeField !== "number" ||
-    !Number.isInteger(magnitudeField) ||
-    magnitudeField < 0
-  ) {
-    throw new Error(
-      "tokenAllowance JSON missing or invalid unitMagnitude (update CLI / ledger-live):\n" + output,
-    );
-  }
-
-  return { allowanceStr, unitMagnitude: magnitudeField };
-}
-
-function setDisableBroadcast(value: string) {
-  const originalDisableTransactionBroadcast = process.env.DISABLE_TRANSACTION_BROADCAST;
-  process.env.DISABLE_TRANSACTION_BROADCAST = value;
-  return originalDisableTransactionBroadcast;
-}
-
-function restoreDisableBroadcast(originalDisableTransactionBroadcast: string | undefined) {
-  if (originalDisableTransactionBroadcast === undefined) {
-    delete process.env.DISABLE_TRANSACTION_BROADCAST;
-  } else {
-    process.env.DISABLE_TRANSACTION_BROADCAST = originalDisableTransactionBroadcast;
-  }
-}
+export const isTokenAllowanceSufficientCommand = createIsTokenAllowanceSufficientCommand(
+  (opts: GetTokenAllowanceOpts) => CLI.getTokenAllowance(opts),
+);
