+++ e2e/mobile/page/trade/deviceValidation.page.ts
@@ -1,8 +1,11 @@
+import CommonPage from "../common.page";
+import { delay } from "../../helpers/commonHelpers";
 import { Step } from "jest-allure2-reporter/api";
 
-export default class DeviceValidationPage {
+export default class DeviceValidationPage extends CommonPage {
   validationScrollViewId = "device-validation-scroll-view";
-  validationAmount = () => getElementById("device-validation-amount");
+  validationAmountId = "device-validation-amount";
+  validationAmount = () => getElementById(this.validationAmountId);
   validationAddress = () => getElementById("device-validation-address");
   validationProvider = () => getElementById("device-validation-provider");
   validationFees = () => getElementById("device-validation-transaction-fee");
@@ -30,5 +33,25 @@ export default class DeviceValidationPage {
   @Step("Expect fees in device validation screen")
   async expectFees(fees: string) {
     await detoxExpect(this.validationFees()).toHaveText(fees);
+  }
+
+  @Step("Wait for swap device validation and retry")
+  async waitDeviceValidationAndRetry() {
+    const WAIT_TIMEOUT = 30000;
+    const CHECK_INTERVAL = 1000;
+    const startTime = Date.now();
+
+    while (Date.now() - startTime < WAIT_TIMEOUT) {
+      if (await IsIdVisible(this.validationAmountId, 100)) {
+        return;
+      }
+      if (await IsIdVisible(this.proceedButtonId, 100)) {
+        await tapById(this.proceedButtonId);
+        return;
+      }
+      await delay(CHECK_INTERVAL);
+    }
+
+    throw new Error(`Device validation timed out after ${WAIT_TIMEOUT / 1000} seconds.`);
   }
 }
