+++ libs/ledger-live-common/src/e2e/index.ts
@@ -90,3 +90,6 @@ export const sanitizeError = (error: unknown): Error => {
 
   return sanitized;
 };
+
+export * from "./cliCommandBuilders";
+export * from "./tokenAllowanceCommands";
